feature: moves some functions to an internal module
