+++ internal/utils/nullable.go
@@ -1,34 +1,10 @@
-package core
+package utils
 
 import (
 	"encoding/json"
 	"reflect"
 	"time"
 )
-
-// PtrBool is a helper routine that returns a pointer to given boolean value.
-func PtrBool(v bool) *bool { return &v }
-
-// PtrInt is a helper routine that returns a pointer to given integer value.
-func PtrInt(v int) *int { return &v }
-
-// PtrInt32 is a helper routine that returns a pointer to given integer value.
-func PtrInt32(v int32) *int32 { return &v }
-
-// PtrInt64 is a helper routine that returns a pointer to given integer value.
-func PtrInt64(v int64) *int64 { return &v }
-
-// PtrFloat32 is a helper routine that returns a pointer to given float value.
-func PtrFloat32(v float32) *float32 { return &v }
-
-// PtrFloat64 is a helper routine that returns a pointer to given float value.
-func PtrFloat64(v float64) *float64 { return &v }
-
-// PtrString is a helper routine that returns a pointer to given string value.
-func PtrString(v string) *string { return &v }
-
-// PtrTime is helper routine that returns a pointer to given Time value.
-func PtrTime(v time.Time) *time.Time { return &v }
 
 type NullableBool struct {
 	value *bool
