+++ internal/utils/strict_decoder.go
@@ -1,0 +1,12 @@
+package utils
+
+import (
+	"bytes"
+	"encoding/json"
+)
+
+func NewStrictDecoder(data []byte) *json.Decoder {
+	dec := json.NewDecoder(bytes.NewBuffer(data))
+	dec.DisallowUnknownFields()
+	return dec
+}
